refactor: restructure project with cli package
