+++ README.md
@@ -5,13 +5,13 @@ A starting CLI application written in Go.
 ## Building
 
 ```bash
-go build
+go build ./cmd/os-bot
 ```
 
 ## Running
 
 ```
-./os-bot -interactive
+./cmd/os-bot/os-bot -interactive
 ```
 
 In interactive mode, type `help` to see available commands.
